Remove use of old-style cast

+++ mdal/mdal_utils.cpp
@@ -47,7 +47,7 @@ size_t MDAL::toSizeT( const std::string &str )
   int i = atoi( str.c_str() );
   if ( i < 0 ) // consistent with atoi return
     i = 0;
-  return i;
+  return static_cast< size_t >( i );
 }
 
 double MDAL::toDouble( const std::string &str )
